Add tests for POSIX file flush, truncate, and delete. (#3139)
b/302715109

(cherry picked from commit 79adf53a067382d8935bb1faa57b5ae422e8e7fc)

+++ starboard/nplb/posix_compliance/posix_file_read_test.cc
@@ -39,7 +39,7 @@ class PosixRead {
 
 class PosixReadAll {
  public:
-  static ssize_t Read(int file, void* data, int size) {
+  static ssize_t Read(int file, void* data, size_t size) {
     return ReadAll(file, data, size);
   }
 };
